+++ apps/cold-graphql/src/backend/entities/postgresql/organization-file.ts
@@ -26,7 +26,7 @@ export enum OrganizationFilesType {
 	TRANSACTION_CERTIFICATE = 'TRANSACTION_CERTIFICATE',
 	SUPPLIER_STATEMENT = 'SUPPLIER_STATEMENT',
 	SUPPLIER_AGREEMENT = 'SUPPLIER_AGREEMENT',
-  SUSTAINABILITY_DATA = 'SUSTAINABILITY_DATA',
+	SUSTAINABILITY_DATA = 'SUSTAINABILITY_DATA',
 }
 
 export enum OrganizationFilesProcessingStatus {
@@ -168,7 +168,7 @@ export class OrganizationFile {
 	  }
 	  return await this.sidecar.beforeReadHook(params);
 	}
-
+	
 	@Hook(HookRegister.AFTER_READ)
 	async afterRead(params: ReadHookParams<typeof OrganizationFile, OrgContext>) {
 	  if(!this.sidecar) {
@@ -176,7 +176,7 @@ export class OrganizationFile {
 	  }
 	  return await this.sidecar.afterReadHook(params);
 	}
-
+	
 	@Hook(HookRegister.BEFORE_UPDATE)
 	async beforeUpdate(params: CreateOrUpdateHookParams<typeof OrganizationFile, OrgContext>) {
 	  if(!this.sidecar) {
@@ -184,7 +184,7 @@ export class OrganizationFile {
 	  }
 	  return await this.sidecar.beforeUpdateHook(params);
 	}
-
+	
 	@Hook(HookRegister.AFTER_UPDATE)
 	async afterUpdate(params: CreateOrUpdateHookParams<typeof OrganizationFile, OrgContext>) {
 	  if(!this.sidecar) {
@@ -192,7 +192,7 @@ export class OrganizationFile {
 	  }
 	  return await this.sidecar.afterUpdateHook(params);
 	}
-
+	
 	@Hook(HookRegister.BEFORE_DELETE)
 	async beforeDelete(params: DeleteHookParams<typeof OrganizationFile, OrgContext>) {
 	  if(!this.sidecar) {
@@ -200,7 +200,7 @@ export class OrganizationFile {
 	  }
 	  return await this.sidecar.beforeDeleteHook(params);
 	}
-
+	
 	@Hook(HookRegister.AFTER_DELETE)
 	async afterDelete(params: DeleteHookParams<typeof OrganizationFile, OrgContext>) {
 	  if(!this.sidecar) {