+++ libs/react/src/components/molecules/dataGrid/muiDataGrid.tsx
@@ -1,31 +1,39 @@
 import { HexColors } from '@coldpbc/themes';
 import { MUIDataGridNoRowsOverlay } from '@coldpbc/components';
 import {
-  DataGrid,
-  DataGridProps, GridCallbackDetails,
+  DataGridPro,
+  DataGridProProps, GridCallbackDetails,
   GridColDef, GridFilterModel,
   GridToolbarColumnsButton,
   GridToolbarContainer,
   GridToolbarExport, GridToolbarProps,
   GridToolbarQuickFilter,
-  GridValidRowModel, ToolbarPropsOverrides,
-} from '@mui/x-data-grid';
+  GridValidRowModel,
+} from '@mui/x-data-grid-pro';
 import { twMerge } from 'tailwind-merge';
-import React, {useEffect} from 'react';
+import React, { useEffect } from 'react';
 import {Box} from "@mui/material";
 import {useAuth0Wrapper} from "@coldpbc/hooks";
 import {addToOrgStorage, getFromOrgStorage} from "@coldpbc/lib";
 
-export interface MUIDataGridProps extends DataGridProps {
+export interface MUIDataGridProps extends DataGridProProps {
   rows: GridValidRowModel[];
   columns: GridColDef[];
   showSearch?: boolean;
   showExport?: boolean;
   showManageColumns?: boolean;
   searchKey?: string; // enabled controlled search
+  ref?: React.RefObject<HTMLDivElement>
 }
 
-const CustomDataGridToolbar = (props: GridToolbarProps & ToolbarPropsOverrides) => {
+interface CustomDataGridToolbarProps extends GridToolbarProps {
+  showSearch?: boolean;
+  showExport?: boolean;
+  showManageColumns?: boolean;
+  // any additional custom props
+}
+
+const CustomDataGridToolbar = (props: CustomDataGridToolbarProps) => {
   const { showSearch, showExport, showManageColumns, quickFilterProps } = props;
   if(!showSearch && !showExport && !showManageColumns) {
     return null;
@@ -77,11 +85,9 @@ export const MuiDataGrid = (props: MUIDataGridProps) => {
   }
 
   return (
-    <DataGrid
+    <DataGridPro
+      pagination
       rowHeight={37}
-      getRowClassName={() => {
-        return 'text-tc-primary cursor-pointer bg-gray-10';
-      }}
       columnHeaderHeight={40}
       autoHeight={true}
       {...props}
@@ -109,6 +115,9 @@ export const MuiDataGrid = (props: MUIDataGridProps) => {
         },
         '& .MuiDataGrid-columnHeader:focus-within': {
           outline: 'none',
+        },
+        '& .MuiDataGrid-row': {
+          cursor: 'pointer',
         },
         ...props.sx,
       }}
@@ -172,7 +181,7 @@ export const MuiDataGrid = (props: MUIDataGridProps) => {
           showSearch,
           showExport,
           showManageColumns,
-        },
+        } as CustomDataGridToolbarProps,
       }}
       className={twMerge('text-tc-primary border-[2px] rounded-[2px] border-gray-30 bg-transparent w-full h-auto', props.className)}
       filterDebounceMs={searchKey ? 500 : props.filterDebounceMs}