+++ libs/react/src/components/pages/uploads/uploadsPage.tsx
@@ -9,7 +9,7 @@ import {
 import { useAuth0Wrapper, useColdContext, useGraphQLSWR } from '@coldpbc/hooks';
 import { UploadsQuery } from '@coldpbc/interfaces';
 import { get } from 'lodash';
-import { GridColDef, GridRenderCellParams, GridTreeNodeWithRender } from '@mui/x-data-grid';
+import { GridColDef, GridRenderCellParams, GridTreeNodeWithRender } from '@mui/x-data-grid-pro';
 import { format } from 'date-fns';
 import React, { ReactNode } from 'react';
 import {